pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = ""
authors = ["K-K-Ju <andrunivermail@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi="^0.115.2"
openai="^1.52.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.0"
count-tokens = "^0.7.0"
aiohttp = "^3.10.10"
httpx = "^0.27.2"
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
